+++ src/Repository/FileRepository.php
@@ -3,16 +3,29 @@
 
 namespace Coffreo\CephOdm\Repository;
 
+use Coffreo\CephOdm\EventListener\QueryTruncatedListener;
 use Coffreo\CephOdm\ResultSet\FileResultSet;
 
 /**
  * AbstractRepositoryWrapper for FileRepository
  */
-class FileRepository extends AbstractRepositoryDecorator
+class FileRepository extends AbstractRepositoryDecorator implements QueryTruncatedListener
 {
-    protected function createResultSet(array $result) : \ArrayObject
+    private $bucketNames = [];
+
+    protected function createResultSet(array $result): \ArrayObject
     {
-        return new FileResultSet($result);
+        $bucketNames = $this->bucketNames;
+        $this->bucketNames = [];
+
+        return new FileResultSet($result, $bucketNames);
     }
 
+    /**
+     * @codeCoverageIgnore
+     */
+    public function queryTruncated(array $bucketNames)
+    {
+        $this->bucketNames = $bucketNames;
+    }
 }